revert: remove repeat controlled fields support
When it comes to setting the value of controlled fields, it's impossible to tell whether an array is supposed to be spread between multiple instances of a controlled field or if it's meant for one instance of the field. I think it should be fine if controlled fields only support array syntax and not repeated fields.